test validate data not size

+++ paimon-spark/paimon-spark-common/src/main/scala/org/apache/paimon/spark/FormatTableStatistics.scala
@@ -35,7 +35,7 @@ case class FormatTableStatistics[T <: PaimonFormatTableBaseScan](scan: T) extend
       .map(
         split => {
           if (split.length() != null) {
-            split.length()
+            split.length().longValue()
           } else {
             split.fileSize()
           }
